% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleData.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example Mendelian randomization genotype summary level dataset}
\format{
\subsection{\code{who}}{

A data frame with 76 rows and 5 columns:
\describe{
\item{SNP}{SNP genotype}
\item{beta.exposure}{SNP-exposure estimates}
\item{beta.outcome}{SNP-outcome estimates}
\item{se.exposure}{Standard errors of the SNP-exposure estimates}
\item{se.outcome}{Standard errors of the SNP-outcome estimates}
}
}
}
\usage{
ExampleData
}
\description{
An example Mendelian randomization dataset.
}
\keyword{datasets}
